# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: M4Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPublic"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
  /users/me:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Chats
...
  /chats/{chatId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a chat
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send message
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    put:
      operationId: editMessage
      tags: [messages]
      summary: Edit message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Message edited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete message (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message deleted
  /messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      summary: Add reaction
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji:
                  type: string
      responses:
        "201":
          description: Reaction added
  /messages/{messageId}/pin:
    post:
      operationId: pinMessage
      tags: [messages]
      summary: Pin/unpin message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pinned:
                  type: boolean
      responses:
        "200":
          description: Pinned state updated
  /messages/search:
    get:
      operationId: searchMessages
      tags: [messages]
      summary: Search messages
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: chatId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  # Files
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
        twoFAEnabled:
          type: boolean
      required:
        - id
        - username
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        isOnline:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
      required:
        - id
        - username
    UpdateProfileRequest:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
...
    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatar:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [text, image, file, voice, video, poll]
        mediaUrl:
          type: string
        replyTo:
          type: integer
        isEdited:
          type: boolean
        isDeleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
      required:
        - id
        - chatId
        - messageType
        - createdAt
...
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        messageType:
          type: string
          default: text
        mediaUrl:
          type: string
        replyTo:
          type: integer
    Reaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        mine:
          type: boolean
...